加锁

+++ douban.py
@@ -8,6 +8,7 @@ import json
 import tempfile
 import os
 import os.path
+import threading
 
 import cookie
 
@@ -27,6 +28,7 @@ class Douban(object):
         self.opener = urllib2.build_opener(cookieHandler)
         self.songs = []
         self.tempfile = None
+        self.lock = threading.Lock()
 
     def _open(self, type='n', sid=None, channel=0, pt=None):
         params = {}
@@ -48,40 +50,45 @@ class Douban(object):
         self.songs = j['song']
         self.songs.reverse()
 
-    def next(self, song = None):
-        if not song:
-            # new
-            if not self.songs:
-                res = self._open()
+    def next(self, song = None, blocking = True):
+        if not self.lock.acquire(blocking):
+            return
+        try:
+            if not song:
+                # new
+                if not self.songs:
+                    res = self._open()
+                    self._parse(res)
+            elif song.time >= song.length:
+                # reach end
+                res = self._open(type='e', sid=song.sid, pt=song.time)
+                res.close()
+                if not self.songs:
+                    res = self._open(type='p', sid=song.sid, pt=0)
+                    self._parse(res)
+            else:
+                # hand
+                res = self._open(type='s', sid=song.sid, pt=song.time)
                 self._parse(res)
-        elif song.time >= song.length:
-            # reach end
-            res = self._open(type='e', sid=song.sid, pt=song.time)
-            res.close()
+
+            json = self.songs.pop()
+            result = Song(json)
+            if self.tempfile:
+                os.remove(self.tempfile)
+            fd, self.tempfile = tempfile.mkstemp()
+            r = self.opener.open(result.url)
+            data = r.read()
+            r.close()
+            os.write(fd, data)
+            os.close(fd)
+            result.file = self.tempfile
+
             if not self.songs:
-                res = self._open(type='p', sid=song.sid, pt=0)
+                res = self._open(type='p', sid=result.sid, pt=0)
                 self._parse(res)
-        else:
-            # hand
-            res = self._open(type='s', sid=song.sid, pt=song.time)
-            self._parse(res)
-
-        json = self.songs.pop()
-        result = Song(json)
-        if self.tempfile:
-            os.remove(self.tempfile)
-        fd, self.tempfile = tempfile.mkstemp()
-        r = self.opener.open(result.url)
-        data = r.read()
-        r.close()
-        os.write(fd, data)
-        os.close(fd)
-        result.file = self.tempfile
-
-        if not self.songs:
-            res = self._open(type='p', sid=result.sid, pt=0)
-            self._parse(res)
-        return result
+            return result
+        finally:
+            self.lock.release()
 
 
     def like(self, song):
